OBJ_LOADER_DEMO: Move material id to its own GBuffer
Since the albedo gbuffer was UNORM, all the material ids were being
clamped to 1. Instead of making the albedo buffer bigger, I chose to
shift the albedo to R11G11B10 and then create a R16 buffer for materialId.
This gives albedo a little bit more precision and gives materialId plenty of room
for lots of materials per frame in the future. The downside is a bit more bandwidth
and 16 bits more storage per sample.