COMMON/OBJ_LOADER_DEMO: Clean up vector push_back memory management
Before I had been doing a push_back with an empty object and then getting a pointer to
the last element of the vector. While this worked, it was hiding some byte alignment issues
as well as not being very clean looking. I opted to either eat a copy construtor or use the
heap. If heap becomes a problem, I can look into a fixed width memory-pool