DOCUMENTATION: Create a json format for representing scene information
While I could have used xml here, I personally like json over xml because it's a
little less verbose and therefore, easier to read at a glance. That said, the lack
of verbosity can mean it's challenging for a newcomer to know what each
field is supposed to be. I added quite a few comments as well as a json schema
in order to help overcome this potential confusion.