OBJ_LOADER_DEMO: Implement geometry instancing for forward shading
Directly inspired by http://www.slideshare.net/DICEStudio/directx-11-rendering-in-battlefield-3
slide 30. The idea is to break any instance data you have into float4's. Then all instance data for the frame
is stored in a giant buffer. This removes the need for individual instance buffers for each model
that needs to be instanced as well as allows for other types of instance data (skinning etc.)