PBR_DEMO: Add crude support for all model types in a scene
I don't really like the idea of casting back and forth to a void pointer. Seems really
hacky. But I wanted to test to make sure all the scene logic was working.
Polymorphism seems like a possibility. Perhaps an abstract function to return a
new Model* in which each child implements?