ENGINE: Create and implement a material cache
It was difficult to choose how to best store and query the cache. Since a material
is made up of two vectors, it becomes non trivial to do comparisons and hashes.
Since material look-ups should happen relatively infrequently, I chose to go for an
unordered_set. Therefore, I implemented operator= and a Hash functor for Material.

To avoid needless copies/moves, the cache takes references to the raw elements
that make up a material, then use those to try to find/create a Material