GRAPHICS: No longer track input layout in draw* commands
Input layout is linked to the vertex shader. ATM, we force a full CommandBucket
flush when a vertex shader is changed. Therefore, it's useless to store/sort
on input layout, since all draw calls within the Bucket will be using the same
vertexshader and in turn, the same input layout.